requirements.txt
git+https://github.com/facebookresearch/pytorch3d.git@stable
git+https://github.com/otaheri/MANO.git
git+https://github.com/otaheri/chamfer_distance.git
git+https://github.com/otaheri/bps_torch.git
git+https://github.com/gpaschalidis/ReachingField.git
git+https://github.com/gpaschalidis/CGrasp.git
git+https://github.com/gpaschalidis/CReach.git
open3d==0.18.0
smplx
tensorboardX
loguru
omegaconf
yacs
torchgeometry
pillow
tqdm
Ninja
trimesh
pyrender
PyYAML
matplotlib==3.8.2
contourpy==1.2.1
numpy==1.22.4
